tests/wpt/web-platform-tests/html/dom/dynamic-markup-insertion/document-writeln/document.writeln-02.html

<!DOCTYPE html>
<title>document.writeln and null/undefined</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-writeln%28%29">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#documents-in-the-dom">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var iframe = document.createElement("iframe");
  document.body.appendChild(iframe);
  doc = iframe.contentDocument;
  test(function() {
    doc.open();
    doc.writeln(null);
    doc.close();
    assert_equals(doc.documentElement.textContent, "null\n");
  }, "document.writeln(null)");
  test(function() {
    doc.open();
    doc.writeln(undefined);
    doc.close();
    assert_equals(doc.documentElement.textContent, "undefined\n");
  }, "document.writeln(undefined)");
}, "Calling document.writeln with null and undefined");
</script>
